requirements.txt
aiofiles==24.1.0
aiogoogle==5.15.0
aiogram==3.20.0.post0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
cachetools==5.5.2
certifi==2025.6.15
click==8.2.1
dnspython==2.7.0
fastapi==0.115.13
frozenlist==1.7.0
google-auth==2.40.3
h11==0.16.0
idna==3.10
magic-filter==1.0.12
motor==3.7.1
multidict==6.5.0
propcache==0.3.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
pymongo==4.13.2
python-dotenv==1.1.0
PyYAML==6.0.2
rsa==4.9.1
sniffio==1.3.1
starlette==0.46.2
tonyg-rfc3339==0.1
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.34.3
yarl==1.20.1
